requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File Processing
python-multipart==0.0.6
pillow==10.1.0
PyMuPDF==1.23.5

# HTTP Client for LM Studio API
requests==2.31.0
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
